+++ app/diograph-store.ts
@@ -71,6 +71,38 @@ export class DiographStore {
     })
   }
 
+  static deleteDiory(id): Promise<any> {
+    return DiographApi.delete(id).then(response => {
+      return null
+    })
+  }
+
+  static getConnection(fromDioryId, toDioryId): Promise<any> {
+    if (fromDioryId === undefined || toDioryId === undefined) { throw "Required two ids not given to DiographStore.getConnection()" }
+    return DiographApi.get([fromDioryId, toDioryId], "connections").then(response => {
+      this.datastore.sync(response)
+      return new Connection(this.datastore.find("connections", response.data[0].id))
+    })
+  }
+
+  static deleteConnection(fromDioryId, toDioryId): Promise<any> {
+    if (fromDioryId === undefined || toDioryId === undefined) { throw "Required two ids not given to DiographStore.deleteConnection()" }
+    return this.getConnection(fromDioryId, toDioryId).then(connection => {
+      return DiographApi.delete(connection.id, "connections").then(response => {
+        return null
+      })
+    })
+  }
+
+  static deleteStrongConnection(fromDioryId, toDioryId): Promise<any> {
+    if (fromDioryId === undefined || toDioryId === undefined) { throw "Required two ids not given to DiographStore.deleteStrongConnection()" }
+    let fromToPromise = this.deleteConnection(fromDioryId, toDioryId)
+    let toFromPromise = this.deleteConnection(toDioryId, fromDioryId)
+    return Promise.all([fromToPromise, toFromPromise]).then(() => {
+      return null
+    })
+  }
+
   static connectDiories(fromDioryId, toDioryId): Promise<ConnectionObject> {
     if (fromDioryId === undefined || toDioryId === undefined) { throw "DiographStore.connectDiories() requires two parameters" }
     let fromDiory, toDiory