+++ spec/diograph-store/diories/get-diory-spec.ts
@@ -1,7 +1,7 @@
 import { DiographStore } from "../../../app/diograph-store"
 import { Diory } from "../../../app/models/diory"
 
-describe("DiographStore .get()", () => {
+describe("DiographStore .get('diories')", () => {
 
   beforeEach(() => {
     DiographStore.setAuthToken("df548369-d0a2-4ca5-b28a-dd4fb14c1f08")