+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.1.0] 2017-11-29
+
+### Added
+- DiographStore
+  - .getConnection(fromDioryId, toDioryId)
+  - .deleteDiory(id)
+  - .deleteConnection(fromDioryId, toDioryId)
+  - .deleteStrongConnection(fromDioryId, toDioryId)
+
+- DiographApi
+  - .get([1, 2], "connections")
+  - .delete(1, "connections")
+
+### Changed
+- tests don't pollute database anymore
+
 ## [0.0.10] 2017-10-31
 
 ### Added